Move db extension methods to kv extensions
