+++ src/Nethermind/Nethermind.Db.Test/ColumnsDbTests.cs
@@ -4,6 +4,7 @@
 using System;
 using System.IO;
 using FluentAssertions;
+using Nethermind.Core;
 using Nethermind.Core.Extensions;
 using Nethermind.Core.Test.Builders;
 using Nethermind.Db.Rocks;
