+++ src/Nethermind/Nethermind.Db.Test/MemDbTests.cs
@@ -3,6 +3,7 @@
 
 using System.Collections.Generic;
 using FluentAssertions;
+using Nethermind.Core;
 using Nethermind.Core.Test.Builders;
 using NUnit.Framework;
